src/livekit-mcp-client_redtius/requirements.txt
livekit-agents
python-dotenv
mcp
RestrictedPython
